% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{annotate_variants}
\alias{annotate_variants}
\title{Annotates variant table}
\usage{
annotate_variants(toannotate, global)
}
\arguments{
\item{toannotate}{an intermediate reduced vcf style dataframe}

\item{global}{internal list with runtime vars}
}
\value{
intermediate data.frame with genome level annotation
}
\description{
Annotates variant table
}
\keyword{internal}
