% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_fasta.R
\name{handle_ambiguous_bases_from_snp_sites}
\alias{handle_ambiguous_bases_from_snp_sites}
\title{Converts ambiguous ( dimer) base calls in the vcf from snp-sites, to 50:50 alt frequency}
\usage{
handle_ambiguous_bases_from_snp_sites(vcf_text)
}
\arguments{
\item{vcf_text}{a text list of the vcf file from snp-sites}
}
\value{
the same vcf text but altered AB's
}
\description{
This is to better support sanger seq fasta files typically used in hospitals
for an ambiguous base (AB), identifies positions with AB, 
then replaces the vcf line with the alt choice. and sets it to 50% frequency
this assumes you do not have a mixture of two alternate alleles
Filters away all over AB
}
\keyword{internal}
