% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_fasta.R
\name{handle_fasta}
\alias{handle_fasta}
\title{Handles fasta files}
\usage{
handle_fasta(dir)
}
\arguments{
\item{dir}{the directory containing}
}
\value{
vcf file and fasta alignment
}
\description{
Decides whether the provided fasta file is, whole genome Merlin aligned/ assembled,
Whole genome other, or only covers a fraction of the genome.
If not WG Merlin type, then it aligns using MAFFT --add --keeplength.
}
\details{
Then converts to VCF using snp-sites
}
\keyword{internal}
