% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herpesdrg_data.R
\name{herpesdrg_data}
\alias{herpesdrg_data}
\title{Returns a dataframe of the herpesdrg database}
\usage{
herpesdrg_data()
}
\value{
dataframe of the herpesdrg database
}
\description{
Returns a dataframe of the herpesdrg database
}
